refactor(phase-E): move mgmt manifest rendering to Provider.RenderMgmtManifest
Extract the 300-line Proxmox-specific management cluster manifest logic
out of internal/capi/pivot/ and into a new Provider interface method.

- Add RenderMgmtManifest to Provider interface; MinStub returns ErrNotApplicable
- New internal/provider/proxmox/mgmt_manifest.go: full implementation
  (clusterctl generate, applyMgmtPatches, role/resource overrides,
  scalarCSVToYAMLList, injectMemoryAdjustment, readAuthorizedKeys)
- internal/capi/pivot/manifest.go: renderManagementManifest shrinks
  to a 4-line provider.For(cfg).RenderMgmtManifest(...) dispatch
- Pivot package is now provider-agnostic (no Proxmox imports)

Closes #13 (Phase E abstraction goal).
