+++ internal/capi/pivot/manifest.go
@@ -7,298 +7,27 @@ import (
 	"crypto/sha256"
 	"encoding/hex"
 	"fmt"
-	"io"
 	"os"
-	"os/exec"
-	"path/filepath"
 	"regexp"
 	"sort"
 	"strings"
 
-	"github.com/lpasquali/yage/internal/capi/manifest"
 	"github.com/lpasquali/yage/internal/config"
 	"github.com/lpasquali/yage/internal/platform/airgap"
 	"github.com/lpasquali/yage/internal/platform/sysinfo"
-	"github.com/lpasquali/yage/internal/ui/logx"
+	"github.com/lpasquali/yage/internal/provider"
 )
 
-// renderManagementManifest generates the CAPI manifest for the management
-// cluster on Proxmox by shelling out to `clusterctl generate cluster`,
-// mirroring capimanifest.GenerateWorkloadManifestIfMissing but with MGMT_*
-// inputs. Writes the rendered manifest to cfg.Mgmt.CAPIManifest (or a temp
-// file when empty) and returns the path on disk.
-//
-// After generation, applies the same four patches the workload manifest
-// gets, except via mgmt-scoped helpers (the workload helpers in
-// capimanifest read cfg.CAPIManifest / cfg.WorkloadCluster*; we wrap them
-// by temporarily redirecting the relevant fields).
-//
-// clusterctlCfgPath is the same ephemeral clusterctl config file the kind
-// init phase generated; reuse it so PROXMOX_URL / token wiring matches.
+// renderManagementManifest dispatches to the active provider's
+// RenderMgmtManifest implementation. All provider-specific logic
+// (clusterctl env vars, manifest patches, template IDs) lives in the
+// provider package — this wrapper keeps the pivot package provider-agnostic.
 func renderManagementManifest(cfg *config.Config, clusterctlCfgPath string) (string, error) {
-	if clusterctlCfgPath == "" {
-		return "", fmt.Errorf("renderManagementManifest: clusterctlCfgPath is empty (call SyncClusterctlConfigFile first)")
-	}
-	if _, err := os.Stat(clusterctlCfgPath); err != nil {
-		return "", fmt.Errorf("renderManagementManifest: clusterctl config %s: %w", clusterctlCfgPath, err)
-	}
-
-	// Determine output path: cfg.Mgmt.CAPIManifest wins; otherwise a
-	// stable temp file under os.TempDir so subsequent runs reuse it.
-	out := cfg.Mgmt.CAPIManifest
-	if out == "" {
-		f, err := os.CreateTemp("", "capi-mgmt-*.yaml")
-		if err != nil {
-			return "", fmt.Errorf("create tmp manifest: %w", err)
-		}
-		out = f.Name()
-		f.Close()
-		cfg.Mgmt.CAPIManifest = out
-	}
-
-	// Reuse an existing non-empty manifest. The mgmt manifest has no
-	// "stale vs config.yaml" check — there is no per-mgmt analogue of
-	// the workload bootstrap config — so we keep it simple: if the
-	// file is present and non-empty, reuse it.
-	if fi, err := os.Stat(out); err == nil && fi.Size() > 0 {
-		logx.Log("Reusing existing management cluster manifest %s.", out)
-		return out, nil
-	}
-
-	logx.Log("Generating management cluster manifest with clusterctl (target: Proxmox)…")
-
-	// Required inputs.
-	var missing []string
-	if cfg.Providers.Proxmox.URL == "" {
-		missing = append(missing, "PROXMOX_URL")
-	}
-	if cfg.Providers.Proxmox.Region == "" {
-		missing = append(missing, "PROXMOX_REGION")
-	}
-	if cfg.Providers.Proxmox.Node == "" {
-		missing = append(missing, "PROXMOX_NODE")
-	}
-	if cfg.Providers.Proxmox.TemplateID == "" {
-		missing = append(missing, "PROXMOX_TEMPLATE_ID")
-	}
-	if cfg.Mgmt.ControlPlaneEndpointIP == "" {
-		missing = append(missing, "MGMT_CONTROL_PLANE_ENDPOINT_IP")
-	}
-	if cfg.Mgmt.NodeIPRanges == "" {
-		missing = append(missing, "MGMT_NODE_IP_RANGES")
-	}
-	if len(missing) > 0 {
-		return "", fmt.Errorf("missing inputs for management manifest: %s", strings.Join(missing, " "))
-	}
-
-	bridge := cfg.Providers.Proxmox.Bridge
-	sourceNode := cfg.Providers.Proxmox.SourceNode
-	if sourceNode == "" {
-		sourceNode = cfg.Providers.Proxmox.Node
-	}
-	sshKeys := cfg.VMSSHKeys
-	if sshKeys == "" {
-		sshKeys = readAuthorizedKeys()
-	}
-
-	// Stage to a tmp file and rename atomically when complete (matches
-	// capimanifest.GenerateWorkloadManifestIfMissing).
-	tmp, err := os.CreateTemp("", "capi-mgmt-gen-*.yaml")
+	p, err := provider.For(cfg)
 	if err != nil {
-		return "", fmt.Errorf("tmp manifest: %w", err)
+		return "", err
 	}
-	tmpPath := tmp.Name()
-	tmp.Close()
-
-	// K3s mode short-circuits clusterctl: render the embedded K3s
-	// template directly to the staging path.
-	if cfg.BootstrapMode == "k3s" {
-		logx.Log("BOOTSTRAP_MODE=k3s — rendering embedded K3s manifest for management cluster (no clusterctl generate).")
-		if err := capimanifest.MaterializeK3sManifest(cfg, true, tmpPath); err != nil {
-			return "", fmt.Errorf("render K3s mgmt manifest: %w", err)
-		}
-		return tmpPath, nil
-	}
-
-	// `clusterctl generate cluster` reads template variables from env;
-	// the workload phase exports WORKLOAD_*-equivalent vars (BOOT_VOLUME_*,
-	// NUM_*, MEMORY_MIB, NODE_IP_RANGES, CONTROL_PLANE_ENDPOINT_IP, etc.).
-	// We override those with MGMT_* values so the same Proxmox provider
-	// template renders the management cluster.
-	args := []string{
-		"generate", "cluster", cfg.Mgmt.ClusterName,
-		"--config", clusterctlCfgPath,
-		"--kubernetes-version", cfg.Mgmt.KubernetesVersion,
-		"--control-plane-machine-count", cfg.Mgmt.ControlPlaneMachineCount,
-		"--worker-machine-count", cfg.Mgmt.WorkerMachineCount,
-		"--infrastructure", cfg.InfraProvider,
-	}
-	cmd := exec.Command("clusterctl", args...)
-	cmd.Env = append(os.Environ(),
-		"PROXMOX_URL="+cfg.Providers.Proxmox.URL,
-		"PROXMOX_REGION="+cfg.Providers.Proxmox.Region,
-		"PROXMOX_NODE="+cfg.Providers.Proxmox.Node,
-		"PROXMOX_TEMPLATE_ID="+cfg.Providers.Proxmox.TemplateID,
-		"PROXMOX_POOL="+cfg.Providers.Proxmox.Mgmt.Pool,
-		"TEMPLATE_VMID="+cfg.Providers.Proxmox.TemplateID,
-		"BRIDGE="+bridge,
-		"PROXMOX_SOURCENODE="+sourceNode,
-		"VM_SSH_KEYS="+sshKeys,
-		"CONTROL_PLANE_ENDPOINT_IP="+cfg.Mgmt.ControlPlaneEndpointIP,
-		"NODE_IP_RANGES="+cfg.Mgmt.NodeIPRanges,
-		"GATEWAY="+cfg.Gateway,
-		"IP_PREFIX="+cfg.IPPrefix,
-		"DNS_SERVERS="+cfg.DNSServers,
-		"ALLOWED_NODES="+cfg.AllowedNodes,
-		"PROXMOX_CLOUDINIT_STORAGE="+cfg.Providers.Proxmox.CloudinitStorage,
-		// MGMT_* hardware sizing maps onto the same template vars the
-		// workload uses (the worker* knobs in cfg). Both control-plane
-		// and worker entries in the rendered manifest get these values;
-		// patchPMTBlockMgmt below rewrites them per-role afterwards.
-		"BOOT_VOLUME_DEVICE="+cfg.Providers.Proxmox.Mgmt.ControlPlaneBootVolumeDevice,
-		"BOOT_VOLUME_SIZE="+cfg.Providers.Proxmox.Mgmt.ControlPlaneBootVolumeSize,
-		"NUM_SOCKETS="+cfg.Providers.Proxmox.Mgmt.ControlPlaneNumSockets,
-		"NUM_CORES="+cfg.Providers.Proxmox.Mgmt.ControlPlaneNumCores,
-		"MEMORY_MIB="+cfg.Providers.Proxmox.Mgmt.ControlPlaneMemoryMiB,
-	)
-
-	outFile, err := os.Create(tmpPath)
-	if err != nil {
-		return "", fmt.Errorf("open tmp manifest: %w", err)
-	}
-	cmd.Stdout = outFile
-	cmd.Stderr = os.Stderr
-	if runErr := cmd.Run(); runErr != nil {
-		outFile.Close()
-		os.Remove(tmpPath)
-		return "", fmt.Errorf("clusterctl generate cluster (mgmt) failed: %w", runErr)
-	}
-	outFile.Close()
-
-	if fi, err := os.Stat(tmpPath); err != nil || fi.Size() == 0 {
-		os.Remove(tmpPath)
-		return "", fmt.Errorf("clusterctl produced an empty mgmt manifest")
-	}
-
-	if err := os.Rename(tmpPath, out); err != nil {
-		// Cross-device rename fallback.
-		src, _ := os.Open(tmpPath)
-		dst, _ := os.Create(out)
-		_, _ = io.Copy(dst, src)
-		src.Close()
-		dst.Close()
-		os.Remove(tmpPath)
-	}
-
-	// --- Apply the four patches the workload manifest gets, but
-	// targeting the mgmt manifest path + mgmt cfg values. The helpers
-	// in internal/capimanifest read cfg.CAPIManifest /
-	// cfg.WorkloadCluster*; redirect those temporarily.
-	if err := applyMgmtPatches(cfg, out); err != nil {
-		return "", fmt.Errorf("apply mgmt patches: %w", err)
-	}
-
-	logx.Log("Generated management cluster manifest %s.", out)
-	return out, nil
-}
-
-// applyMgmtPatches runs the same 4 patches the workload manifest gets
-// (apply_role_resource_overrides, patch_csi_topology_labels,
-// patch_kubeadm_skip_kube_proxy_for_cilium,
-// patch_proxmox_machine_template_spec_revisions) but against the
-// management-cluster manifest at manifestPath, with MGMT_* hardware
-// values taking precedence.
-//
-// The capimanifest helpers operate on cfg.CAPIManifest globally, so we
-// shadow the relevant fields for the duration of the call. Role-resource
-// overrides are re-implemented locally because the shared helper hard-
-// codes the workload cp/worker config and we need to substitute
-// MGMT_CONTROL_PLANE_* for control-plane PMTs.
-func applyMgmtPatches(cfg *config.Config, manifestPath string) error {
-	// 1) Role/resource overrides — locally implemented (mgmt sizing).
-	if err := mgmtRoleResourceOverrides(cfg, manifestPath); err != nil {
-		return fmt.Errorf("role resource overrides: %w", err)
-	}
-
-	// Save + redirect cfg fields the capimanifest helpers read.
-	saveManifest := cfg.CAPIManifest
-	cfg.CAPIManifest = manifestPath
-	defer func() { cfg.CAPIManifest = saveManifest }()
-
-	// 2) CSI topology labels — same region/zone wiring as workload.
-	if err := capimanifest.PatchProxmoxCSITopologyLabels(cfg); err != nil {
-		return fmt.Errorf("csi topology labels: %w", err)
-	}
-
-	// 3) skip addon/kube-proxy for Cilium when KPR is on.
-	if err := capimanifest.PatchKubeadmSkipKubeProxyForCilium(cfg); err != nil {
-		return fmt.Errorf("kubeadm skip kube-proxy: %w", err)
-	}
-
-	// 4) ProxmoxMachineTemplate spec-revision rename.
-	if _, err := capimanifest.PatchProxmoxMachineTemplateSpecRevisions(cfg); err != nil {
-		return fmt.Errorf("PMT spec revisions: %w", err)
-	}
-	return nil
-}
-
-// mgmtRoleResourceOverrides is the mgmt-cluster analogue of
-// capimanifest.ApplyRoleResourceOverrides. We re-implement it here
-// instead of borrowing because the workload helper hard-codes the
-// workload control-plane vs. worker sizing; the management cluster
-// uses MGMT_CONTROL_PLANE_* for both the (typically single) control
-// plane node and any management workers.
-func mgmtRoleResourceOverrides(cfg *config.Config, manifestPath string) error {
-	raw, err := os.ReadFile(manifestPath)
-	if err != nil {
-		return err
-	}
-	text := string(raw)
-
-	// Patch every PMT (control-plane or worker) using the mgmt sizing
-	// and per-role template overrides (with PROXMOX_TEMPLATE_ID as the
-	// catch-all fallback when MGMT_*_TEMPLATE_ID is unset).
-	mgmtCPTpl := firstNonEmptyMgmt(cfg.Providers.Proxmox.Mgmt.ControlPlaneTemplateID, cfg.Providers.Proxmox.TemplateID)
-	mgmtWkTpl := firstNonEmptyMgmt(cfg.Providers.Proxmox.Mgmt.WorkerTemplateID, cfg.Providers.Proxmox.TemplateID)
-	parts := strings.Split(text, "\n---\n")
-	nameRE := regexp.MustCompile(`(?m)^  name:\s*(\S+)\s*$`)
-	for i, doc := range parts {
-		if !strings.Contains(doc, "kind: ProxmoxMachineTemplate") {
-			continue
-		}
-		m := nameRE.FindStringSubmatch(doc)
-		if m == nil {
-			continue
-		}
-		// Replace the five hardware fields. Same regex shapes as the
-		// workload helper (capimanifest.patchPMTBlock).
-		doc = replaceFirstPerLine(doc, `(disk:\s*)[^\n]+`, "${1}"+cfg.Providers.Proxmox.Mgmt.ControlPlaneBootVolumeDevice)
-		doc = replaceFirstPerLine(doc, `(sizeGb:\s*)[^\n]+`, "${1}"+cfg.Providers.Proxmox.Mgmt.ControlPlaneBootVolumeSize)
-		doc = replaceFirstPerLine(doc, `(numSockets:\s*)[^\n]+`, "${1}"+cfg.Providers.Proxmox.Mgmt.ControlPlaneNumSockets)
-		doc = replaceFirstPerLine(doc, `(numCores:\s*)[^\n]+`, "${1}"+cfg.Providers.Proxmox.Mgmt.ControlPlaneNumCores)
-		doc = replaceFirstPerLine(doc, `(memoryMiB:\s*)[^\n]+`, "${1}"+cfg.Providers.Proxmox.Mgmt.ControlPlaneMemoryMiB)
-		// Pick the per-role template override; default to CP for any
-		// PMT whose name doesn't contain "worker" (mgmt typically has 0
-		// workers, but we honor the override if present).
-		tpl := mgmtCPTpl
-		if strings.Contains(m[1], "worker") {
-			tpl = mgmtWkTpl
-		}
-		if tpl != "" {
-			doc = replaceFirstPerLine(doc, `(templateID:\s*)[^\n]+`, "${1}"+tpl)
-		}
-		parts[i] = doc
-	}
-	text = strings.Join(parts, "\n---\n")
-
-	// Convert comma-scalar fields to YAML lists (allowedNodes, dnsServers,
-	// addresses) — same shape as the workload helper.
-	text = scalarCSVToYAMLList(text, "allowedNodes")
-	text = scalarCSVToYAMLList(text, "dnsServers")
-	text = scalarCSVToYAMLList(text, "addresses")
-	text = injectMemoryAdjustment(text, cfg.Providers.Proxmox.MemoryAdjustment)
-
-	return os.WriteFile(manifestPath, []byte(text), 0o644)
+	return p.RenderMgmtManifest(cfg, clusterctlCfgPath)
 }
 
 // renderMgmtCiliumHelmChartProxy renders a HelmChartProxy YAML doc that
@@ -535,91 +264,6 @@ func hexDigit(r rune) int {
 	return 0
 }
 
-// --- small helpers (copies from internal/capimanifest, scoped to pivot) ---
-
-// firstNonEmptyMgmt returns the first non-empty argument; used to pick
-// the per-role mgmt template override with a PROXMOX_TEMPLATE_ID
-// fallback. Local copy so this package doesn't depend on capimanifest.
-func firstNonEmptyMgmt(vals ...string) string {
-	for _, v := range vals {
-		if v != "" {
-			return v
-		}
-	}
-	return ""
-}
-
-func replaceFirstPerLine(s, pattern, repl string) string {
-	re := regexp.MustCompile(pattern)
-	return re.ReplaceAllString(s, repl)
-}
-
-func scalarCSVToYAMLList(text, key string) string {
-	pat := regexp.MustCompile(`(?m)^( *)(` + regexp.QuoteMeta(key) + `):\s*"?([^"\n\[]+)"?\s*$`)
-	return pat.ReplaceAllStringFunc(text, func(line string) string {
-		m := pat.FindStringSubmatch(line)
-		if m == nil {
-			return line
-		}
-		indent := m[1]
-		raw := strings.TrimSpace(m[3])
-		raw = strings.Trim(raw, `"'`)
-		var items []string
-		for _, v := range strings.Split(raw, ",") {
-			v = strings.TrimSpace(v)
-			if v != "" {
-				items = append(items, v)
-			}
-		}
-		lines := []string{indent + key + ":"}
-		for _, it := range items {
-			lines = append(lines, indent+"- "+it)
-		}
-		return strings.Join(lines, "\n")
-	})
-}
-
-var (
-	reKindProxmoxCluster = regexp.MustCompile(`(?m)^kind:\s*ProxmoxCluster\s*$`)
-	reAPIInfraProxmox    = regexp.MustCompile(`(?m)^apiVersion:\s*infrastructure\.cluster\.x-k8s\.io/`)
-	reSchedulerHints     = regexp.MustCompile(`(?m)^  schedulerHints:`)
-	reSpecBlock          = regexp.MustCompile(`(?m)^(spec:\n(?:  .*\n)+)`)
-)
-
-func injectMemoryAdjustment(text, mem string) string {
-	if mem == "" {
-		return text
-	}
-	parts := strings.Split(text, "\n---\n")
-	for i, doc := range parts {
-		if !reKindProxmoxCluster.MatchString(doc) {
-			continue
-		}
-		if !reAPIInfraProxmox.MatchString(doc) {
-			continue
-		}
-		if reSchedulerHints.MatchString(doc) {
-			break
-		}
-		// \n---\n splitting strips the trailing \n from the last line of each
-		// doc, so reSpecBlock's (?:  .*\n)+ misses it and leaves it dangling
-		// after schedulerHints at the wrong indentation. Add it back.
-		if !strings.HasSuffix(doc, "\n") {
-			doc += "\n"
-		}
-		loc := reSpecBlock.FindStringIndex(doc)
-		if loc == nil {
-			break
-		}
-		oldBlock := doc[loc[0]:loc[1]]
-		newBlock := strings.TrimRight(oldBlock, "\n") +
-			"\n  schedulerHints:\n    memoryAdjustment: " + mem + "\n"
-		parts[i] = doc[:loc[0]] + newBlock + doc[loc[1]:]
-		break
-	}
-	return strings.Join(parts, "\n---\n")
-}
-
 func sliceHasLine(s []string, want string) bool {
 	for _, x := range s {
 		if strings.TrimSpace(x) == want {
@@ -643,23 +287,6 @@ func insertAtIndex(s []string, i int, v string) []string {
 	copy(s[i+1:], s[i:])
 	s[i] = v
 	return s
-}
-
-func readAuthorizedKeys() string {
-	home, _ := os.UserHomeDir()
-	raw, err := os.ReadFile(filepath.Join(home, ".ssh", "authorized_keys"))
-	if err != nil {
-		return ""
-	}
-	var keys []string
-	for _, line := range strings.Split(string(raw), "\n") {
-		s := strings.TrimSpace(line)
-		if s == "" || strings.HasPrefix(s, "#") {
-			continue
-		}
-		keys = append(keys, s)
-	}
-	return strings.Join(keys, ",")
 }
 
 // sortedStrings is unused at call-time but kept as a helper so VerifyParity
